requirements.txt
discord.py
fuzzywuzzy
python-Levenshtein